Fixed bugs after teleport
Teleporting leads to a corrupted entity list, which we need to flush before using it again. A dedicated thread handles this now. Also I included a small hack that fixes bugs if the player ports during the attack of a stone. This is a small issue but it was annoying me. Also I moved some more code from the dllmain to the game files.